brace_kinova/requirements.txt
gymnasium
stable-baselines3
torch
numpy
scipy
pyyaml
tensorboard
matplotlib
pygame
